+++ KNN/include/CommonFunc.h
@@ -30,7 +30,7 @@ enum KernelNames :int
 struct MinqueOptions
 {
 	int iterate = 200;
-	double tolerance = 1e-6;
+	float tolerance = 1e-6;
 	int MatrixDecomposition = 0;
 	int altMatrixDecomposition = 3;
 	bool allowPseudoInverse = true;
@@ -39,33 +39,33 @@ struct MinqueOptions
 struct KernelData
 {
 	boost::bimap<int, std::string> fid_iid;
-	Eigen::MatrixXd kernelMatrix;
-	Eigen::MatrixXd VariantCountMatrix;
+	Eigen::MatrixXf kernelMatrix;
+	Eigen::MatrixXf VariantCountMatrix;
 };
 
 struct PhenoData
 {
 	boost::bimap<int, std::string> fid_iid;
-	Eigen::VectorXd Phenotype;
+	Eigen::VectorXf Phenotype;
 };
 
 struct GenoData
 {
 	std::map<int, std::string> fid_iid;
-	Eigen::MatrixXd Geno;               //individual mode, row: individual; col: SNP;
+	Eigen::MatrixXf Geno;               //individual mode, row: individual; col: SNP;
 
 };
 
 
-int Inverse(Eigen::MatrixXd & Ori_Matrix,int DecompositionMode, int AltDecompositionMode, bool allowPseudoInverse);
-double Variance(Eigen::VectorXd &Y);
-double mean(Eigen::VectorXd &Y);
-double isNum(std::string line);
+int Inverse(Eigen::MatrixXf & Ori_Matrix,int DecompositionMode, int AltDecompositionMode, bool allowPseudoInverse);
+float Variance(Eigen::VectorXf &Y);
+float mean(Eigen::VectorXf &Y);
+float isNum(std::string line);
 std::string GetBaseName(std::string pathname);
 std::string GetParentPath(std::string pathname);
-void stripSameCol(Eigen::MatrixXd &Geno);
-void stdSNPmv(Eigen::MatrixXd &Geno);
+void stripSameCol(Eigen::MatrixXf &Geno);
+void stdSNPmv(Eigen::MatrixXf &Geno);
 void set_difference(boost::bimap<int, std::string> &map1, boost::bimap<int, std::string> &map2, std::vector<std::string> &overlap);
-void GetSubMatrix(Eigen::MatrixXd &oMatrix, Eigen::MatrixXd &subMatrix, std::vector<int> rowIds, std::vector<int> colIDs);
-void GetSubVector(Eigen::VectorXd &oVector, Eigen::VectorXd &subVector, std::vector<int> IDs);
+void GetSubMatrix(Eigen::MatrixXf &oMatrix, Eigen::MatrixXf &subMatrix, std::vector<int> rowIds, std::vector<int> colIDs);
+void GetSubVector(Eigen::VectorXf &oVector, Eigen::VectorXf &subVector, std::vector<int> IDs);
