+++ KNN/include/ToolKit.h
@@ -9,18 +9,18 @@ using namespace std;
 class ToolKit
 {
 public:
-	static void ArraytoVector(double ** a, int n, int m, vector<vector<double>>& v, bool Transpose = false);
-	static void Array2toArrat1(double **a, int n, int m, double *b, bool Colfirst=true);
-	static void Vector2toArray1(vector<vector<double>>& v, double *b, bool Colfirst = true);
+	static void ArraytoVector(float ** a, int n, int m, vector<vector<float>>& v, bool Transpose = false);
+	static void Array2toArrat1(float **a, int n, int m, float *b, bool Colfirst=true);
+	static void Vector2toArray1(vector<vector<float>>& v, float *b, bool Colfirst = true);
 	static void Stringsplit(string &org, vector<string> & splited, string delim);
 	static void dec2bin(int num, int *bin);
-	static bool Inv_Cholesky(Eigen::MatrixXd & Ori_Matrix);
-	static bool Inv_LU(Eigen::MatrixXd & Ori_Matrix, Eigen::MatrixXd & Inv_Matrix);
-	static bool Inv_SVD(Eigen::MatrixXd & Ori_Matrix,  bool allowPseudoInverse);
-	static bool Inv_QR(Eigen::MatrixXd & Ori_Matrix,  bool allowPseudoInverse);
-	static bool comput_inverse_logdet_LDLT_mkl(Eigen::MatrixXd &Vi);
-	static bool comput_inverse_logdet_LU_mkl(Eigen::MatrixXd &Vi);
-	static bool comput_inverse_logdet_QR_mkl(Eigen::MatrixXd& Vi);
-	static bool comput_inverse_logdet_SVD_mkl(Eigen::MatrixXd& Vi);
+	static bool Inv_Cholesky(Eigen::MatrixXf & Ori_Matrix);
+	static bool Inv_LU(Eigen::MatrixXf & Ori_Matrix, Eigen::MatrixXf & Inv_Matrix);
+	static bool Inv_SVD(Eigen::MatrixXf & Ori_Matrix,  bool allowPseudoInverse);
+	static bool Inv_QR(Eigen::MatrixXf & Ori_Matrix,  bool allowPseudoInverse);
+	static bool comput_inverse_logdet_LDLT_mkl(Eigen::MatrixXf &Vi);
+	static bool comput_inverse_logdet_LU_mkl(Eigen::MatrixXf &Vi);
+	static bool comput_inverse_logdet_QR_mkl(Eigen::MatrixXf& Vi);
+	static bool comput_inverse_logdet_SVD_mkl(Eigen::MatrixXf& Vi);
 };
