+++ KNN/include/imnq.h
@@ -18,7 +18,7 @@ public:
 	int getIterateTimes();
 	void isEcho(bool isecho);
 private:
-	double tol = 1e-5; //convergence tolerence (def=1e-5)
+	float tol = 1e-5; //convergence tolerence (def=1e-5)
 	int itr = 20;	//iterations allowed (def=20)
 	int Decomposition = Cholesky;
 	int altDecomposition = QR;
@@ -26,7 +26,7 @@ private:
 	int initIterate = 0;
 	bool isecho=true;
 private:
-	Eigen::VectorXd initVCS(); //initialize variance components
+	Eigen::VectorXf initVCS(); //initialize variance components
 	void Iterate();
 
 };
